refactor(editor): simplify environment plugin contract
